Aula 20/08/2024

+++ app/src/main/java/application/controller/AlunoController.java
@@ -3,11 +3,13 @@ package application.controller;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -33,7 +35,13 @@ public class AlunoController {
 
     @GetMapping("/{id}")
     public Aluno details(@PathVariable long id) {
-        return alunoRepo.findById(id).get();
+        Optional<Aluno> resultado = alunoRepo.findById(id);
+        if(resultado.isEmpty()) {
+            throw new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "Aluno Não Encontrado"
+            );
+        }
+        return resultado.get();
     }
 
     @PutMapping("/{id}")
@@ -42,6 +50,18 @@ public class AlunoController {
         @RequestBody Aluno novosDados) {
         Optional<Aluno> resultado = alunoRepo.findById(id);
 
+        if(resultado.isEmpty()) {
+            throw new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "Aluno Não Encontrado"
+            );
+        }
+
+        if(novosDados.getNome().isEmpty()){
+            throw new ResponseStatusException(
+                HttpStatus.BAD_REQUEST, "Nome de Aluno Inválido"
+            );
+        }
+
         resultado.get().setNome(novosDados.getNome());
 
         return alunoRepo.save(resultado.get());
@@ -49,6 +69,12 @@ public class AlunoController {
 
     @DeleteMapping("/{id}")
     public void delete(@PathVariable long id) {
+        if(!alunoRepo.existsById(id)) {
+            throw new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "Aluno Não Encontrado"
+            );
+        }
+
         alunoRepo.deleteById(id);
     }
 }
